Rename helpers->utils in chain unit tests
