+++ framework/src/modules/chain/loader.js
@@ -42,7 +42,7 @@ __private.retries = 5;
  * @memberof modules
  * @see Parent: {@link modules}
  * @requires async
- * @requires helpers/jobs_queue
+ * @requires utils/jobs_queue
  * @requires logic/peer
  * @param {function} cb - Callback function
  * @param {scope} scope - App instance
