+++ framework/src/modules/chain/submodules/peers.js
@@ -35,7 +35,7 @@ const MAX_PEERS = 100;
  * @requires pg-promise
  * @requires semver
  * @requires api/ws/rpc/failure_codes
- * @requires helpers/jobs_queue
+ * @requires utils/jobs_queue
  * @requires logic/peer
  * @param {function} cb - Callback function
  * @param {scope} scope - App instance
