+++ framework/src/modules/chain/utils/newrelic_lisk.js
@@ -27,8 +27,8 @@ newrelicLisk.instrumentBackgroundJobs();
 // TOFIX: fix callbackMethods converted to async in #2579
 // callBackMethods array only support one level of nesting
 const modulesToInstrument = {
-	'./helpers/sequence.js': {
-		identifier: 'helpers.sequence',
+	'./utils/sequence.js': {
+		identifier: 'utils.sequence',
 		callbackMethods: ['add'],
 	},
 	'../submodules/blocks': {
