+++ framework/test/mocha/unit/modules/chain/utils/regexp_tester.js
